major patch (#82)
