+++ PRJ/HopeLine.DataAccess/Interfaces/IRepository.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace HopeLine.DataAccess.Interfaces
 {
@@ -6,7 +7,6 @@ namespace HopeLine.DataAccess.Interfaces
     //TODO : add implementation
     public interface IRepository<T>
     {
-
         IEnumerable<T> GetAll(string include = null);
         void Insert(T obj);
         void Update(T obj);
@@ -15,5 +15,9 @@ namespace HopeLine.DataAccess.Interfaces
         void Delete(T obj);
 
         void Remove(object id);
+
+        void Save();
+
+        Task SaveAsync();
     }
 }
