--- PRJ/HopeLine.DataAccess/Interfaces/IUnitOfWork.cs
@@ -1,10 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace HopeLine.DataAccess.Interfaces
-{
-    public interface IUnitOfWork
-    {
-        void Save();
-        Task SaveAsync();
-    }
-}
