+++ PRJ/HopeLine.Service/Configurations/ConfigureServiceExtension.cs
@@ -1,21 +1,15 @@
 ﻿using HopeLine.DataAccess.DatabaseContexts;
 using HopeLine.DataAccess.Entities;
-using HopeLine.DataAccess.Entities.Base;
 using HopeLine.DataAccess.Interfaces;
 using HopeLine.DataAccess.Repositories;
 using HopeLine.Service.CoreServices;
 using HopeLine.Service.Interfaces;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.IdentityModel.Tokens;
 using Pomelo.EntityFrameworkCore.MySql.Infrastructure;
 using System;
-using System.IdentityModel.Tokens.Jwt;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace HopeLine.Service.Configurations
 {
@@ -32,15 +26,7 @@ namespace HopeLine.Service.Configurations
             services.AddDbContext<ChatDbContext>(opt =>
                             opt.UseInMemoryDatabase("chatdb"));
 
-            services.AddDbContext<ResourcesDbContext>(opt => opt
-                    .UseMySql(APIConstant.ConnectionString,
-                        mysqlOptions =>
-                        {
-                            mysqlOptions
-                                .ServerVersion(new Version(3, 23), ServerType.MySql);
-                        }));
-            //.UseInMemoryDatabase("chatdb"));
-            // .UseSqlServer(APIConstant.ConnectionString));
+          
             services.AddDbContext<HopeLineDbContext>(opt => opt
                                                     .UseMySql(APIConstant.ConnectionString,
                         mysqlOptions =>
@@ -67,12 +53,11 @@ namespace HopeLine.Service.Configurations
 
             //all interface and implementation
             services.AddTransient(typeof(IRepository<>), typeof(Repository<>));
-            services.AddTransient<IRepository<HopeLineUser>, UserRepository>();
             services.AddTransient<IUserService, UserService>();
             services.AddTransient<ICommunication, CommunicationService>();
             services.AddTransient<IMessage, MessageService>();
             services.AddTransient<ICommonResource, CommonResourceService>();
-            services.AddTransient<IUnitOfWork, UnitOfWork>();
+            services.AddTransient<IRepository<HopeLineUser>, UserRepository>();
         }
 
         /// <summary>
@@ -85,7 +70,9 @@ namespace HopeLine.Service.Configurations
             using (var scope = app.ApplicationServices.CreateScope())
             {
                 using (var context = scope.ServiceProvider.GetRequiredService<HopeLineDbContext>())
+                {
                     context.Database.EnsureCreated();
+                }
 
                 //TODO : do populate data HERE!
             }
